move location search into create code form
